systemtest/requirements.txt
pyserial==3.5
asyncio==3.4.3
pytest-asyncio==0.25.3
pynrfjprog==10.24.2
bumble
